업데이트 ph값 업데이트 가능하게 수정 , 공정추적 상태 업데이트 수정

+++ src/main/java/com/org/qualitycore/standardinformation/model/dto/MashingProcessDTO.java
@@ -32,10 +32,10 @@ public class MashingProcessDTO {
     private Double phValue;
 
     @Schema(description = "곡물 비율", example = "1")
-    private Integer grainRatio;
+    private Double grainRatio;
 
     @Schema(description = "물 비율" , example = "4")
-    private Integer waterRatio;
+    private Double waterRatio;
 
     @Schema(description = "물 투입량", example = "3200")
     private Integer waterInputVolume;
